+++ tvb_epilepsy/io/csv.py
@@ -1,7 +1,7 @@
 # encoding=utf8
 
 import numpy as np
-from tvb_epilepsy.io.rdump import rdump
+from tvb_epilepsy.io.r_file_io import rdump
 from tvb_epilepsy.base.utils.log_error_utils import initialize_logger
