+++ tvb_epilepsy/service/model_inversion/stan/stan_service.py
@@ -10,7 +10,7 @@ from tvb_epilepsy.base.utils.log_error_utils import initialize_logger, warning, 
                                                     raise_value_error, raise_not_implemented_error
 from tvb_epilepsy.base.utils.data_structures_utils import isequal_string, ensure_list, sort_dict,  \
                                                     list_of_dicts_to_dicts_of_ndarrays, switch_levels_of_dicts_of_dicts
-from tvb_epilepsy.io.rdump import rdump, rload
+from tvb_epilepsy.io.r_file_io import rdump, rload
 from tvb_epilepsy.io.csv import parse_csv
 from tvb_epilepsy.io.h5_reader import H5Reader
 from tvb_epilepsy.io.h5_writer import H5Writer
